+++ api/src/main/java/io/strimzi/api/ResourceAnnotations.java
@@ -102,4 +102,9 @@ public class ResourceAnnotations {
      */
     public static final String ANNO_STRIMZI_IO_FORCE_REPLACE = STRIMZI_DOMAIN + "force-replace";
 
+    /**
+     * Annotation used to skip the check on broker scale-down
+     */
+    public static final String ANNO_STRIMZI_IO_SKIP_BROKER_SCALEDOWN_CHECK = STRIMZI_DOMAIN + "skip-broker-scaledown-check";
+
 }