+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/model/KafkaCluster.java
@@ -368,6 +368,21 @@ public class KafkaCluster extends AbstractModel implements SupportsMetrics, Supp
     }
 
     /**
+     * Generates list of references to Kafka node ids going to be removed from the Kafka cluster.
+     *
+     * @return  Set of Kafka node ids which are going to be removed
+     */
+    public Set<Integer> removedNodes() {
+        Set<Integer> nodes = new LinkedHashSet<>();
+
+        for (KafkaPool pool : nodePools)    {
+            nodes.addAll(pool.scaledDownNodes());
+        }
+
+        return nodes;
+    }
+
+    /**
      * Generates list of references to Kafka nodes for this Kafka cluster which have the broker role. The references
      * contain both the pod name and the ID of the Kafka node.
      *