+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Sign containers using `cosign`
 * Generate and publish Software Bill of Materials (SBOMs) of Strimzi containers
 * Add support for stopping connectors according to [Strimzi Proposal #54](https://github.com/strimzi/proposals/blob/main/054-stopping-kafka-connect-connectors.md)
+* Make sure brokers are empty before scaling them down
 
 ### Changes, deprecations and removals